define mode1 baseline properly
